React: organize generated files into directories
