+++ generators/reactjs/reactjs_app_base/templates/src/components/Loader.js
