+++ generators/reactjs/reactjs_app_base/templates/src/home/home.js
