+++ generators/reactjs/reactjs_auth/index.js
@@ -2,7 +2,9 @@ const Generator = require('../../util/generator');
 
 module.exports = class ReactJsAuth extends Generator {
   async write() {
-    const dest = this.options.build.dest.client.src
+    const dest = this.options.build.dest.client.src + 'auth/';
+
+    await this.ensureDir(dest);
 
     await this.copyTemplate(
       this.templatePath(__dirname, 'Login.js'),
